# GymZZC/Backend/src/main/java/com/backend/gymsync/controller/RutinaController.java
package com.backend.gymsync.controller;

import com.backend.gymsync.entity.Rutina;
import com.backend.gymsync.entity.Usuario;
import com.backend.gymsync.entity.Progreso;
import com.backend.gymsync.service.interfaces.RutinaServiceInterface;

import org.springframework.http.HttpStatus;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import jakarta.validation.Valid;
import java.util.List;

@RestController
@RequestMapping("/api/rutinas")
@CrossOrigin(origins = "*")
public class RutinaController {

    @Autowired
    private RutinaServiceInterface rutinaService;

    @GetMapping
    public ResponseEntity<List<Rutina>> getAllRutinas() {
        return ResponseEntity.ok(rutinaService.findAll());
    }

    @GetMapping("/{id}")
    public ResponseEntity<Rutina> getRutinaById(@PathVariable Integer id) {
        return rutinaService.findById(id)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @GetMapping("/cliente/{clienteId}")
    public ResponseEntity<List<Rutina>> getRutinasByCliente(@PathVariable Integer clienteId) {
        return ResponseEntity.ok(rutinaService.findByClienteId(clienteId));
    }

    @GetMapping("/entrenador/{entrenadorId}")
    public ResponseEntity<List<Rutina>> getRutinasByEntrenador(@PathVariable Integer entrenadorId) {
        return ResponseEntity.ok(rutinaService.findByEntrenadorId(entrenadorId));
    }

    // Nuevo endpoint para obtener todos los progresos de una rutina
    @GetMapping("/{id}/progresos")
    public ResponseEntity<List<Progreso>> getProgresosByRutina(@PathVariable Integer id) {
        return rutinaService.findById(id)
                .map(rutina -> ResponseEntity.ok(rutina.obtenerTodosLosProgresos()))
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public ResponseEntity<Rutina> crearRutina(@Valid @RequestBody Rutina rutina) {
        // Validar que el entrenador tenga rol correcto
        if(!rutina.getEntrenador().getRol().equals(Usuario.Rol.ENTRENADOR)) {
            return ResponseEntity.badRequest().build();
            // return ResponseEntity.badRequest().body("El entrenador no es válido");
        }

        Rutina savedRutina = rutinaService.save(rutina);
        return ResponseEntity.ok(savedRutina);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Rutina> actualizarRutina(@PathVariable Integer id, @Valid @RequestBody Rutina rutina) {
        
        if (!rutinaService.existsById(id)) {
            return ResponseEntity.notFound().build();
        }
        
        rutina.setId(id);
        return ResponseEntity.ok(rutinaService.save(rutina));
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> eliminarRutina(@PathVariable Integer id) {
        if (!rutinaService.findById(id).isPresent()) {
            return ResponseEntity.notFound().build();
        }
        
        try {
            rutinaService.deleteById(id);
            return ResponseEntity.noContent().build();
        } catch (Exception e) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
        }
    }
}