# GymZZC/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Entrenamiento
  version: 1.0.0
  description: API para gestionar usuarios, rutinas, ejercicios y progreso de entrenamiento
servers:
  - url: https://api.tuentrenador.com/v1
    description: Servidor de producción
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
paths:
...
  /rutinas:
    get:
      summary: Obtiene todas las rutinas
      responses:
        '200':
          description: Lista de rutinas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rutina'
    post:
      summary: Crea una nueva rutina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RutinaInput'
      responses:
        '201':
          description: Rutina creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rutina'
  /rutinas/{id}:
    get:
      summary: Obtiene una rutina por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles de la rutina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rutina'
    put:
      summary: Actualiza una rutina existente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RutinaInput'
      responses:
        '200':
          description: Rutina actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rutina'
    delete:
      summary: Elimina una rutina
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Rutina eliminada
  /rutinas/cliente/{cliente_id}:
    get:
      summary: Obtiene rutinas asignadas a un cliente
      parameters:
        - in: path
          name: cliente_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de rutinas del cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rutina'
  /rutinas/entrenador/{entrenador_id}:
    get:
      summary: Obtiene rutinas creadas por un entrenador
      parameters:
        - in: path
          name: entrenador_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de rutinas del entrenador
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rutina'

  # =============================================
  # ENDPOINTS DE EJERCICIOS
  # =============================================
  /ejercicios:
    get:
      summary: Obtiene todos los ejercicios
      responses:
        '200':
          description: Lista de ejercicios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ejercicio'
    post:
      summary: Crea un nuevo ejercicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EjercicioInput'
      responses:
        '201':
          description: Ejercicio creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ejercicio'
  /ejercicios/{id}:
    get:
      summary: Obtiene un ejercicio por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del ejercicio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ejercicio'
    put:
      summary: Actualiza un ejercicio existente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EjercicioInput'
      responses:
        '200':
          description: Ejercicio actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ejercicio'
    delete:
      summary: Elimina un ejercicio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ejercicio eliminado

  # =============================================
  # GESTIÓN DE EJERCICIOS EN RUTINAS
  # =============================================
...
  /rutinas/{rutina_id}/ejercicios/{ejercicio_id}:
    get:
      summary: Obtiene detalles de un ejercicio en una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del ejercicio en la rutina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutinaEjercicio'
    put:
      summary: Modifica un ejercicio en una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RutinaEjercicioInput'
      responses:
        '200':
          description: Ejercicio actualizado en la rutina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutinaEjercicio'
    delete:
      summary: Quita un ejercicio de una rutina
      parameters:
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ejercicio eliminado de la rutina

  # =============================================
  # ENDPOINTS DE PROGRESO
  # =============================================
  /progreso:
    get:
      summary: Obtiene todos los registros de progreso
      responses:
        '200':
          description: Lista de progresos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progreso'
    post:
      summary: Registra nuevo progreso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgresoInput'
      responses:
        '201':
          description: Progreso registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progreso'
  /progreso/{id}:
    get:
      summary: Obtiene un progreso por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del progreso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progreso'
    put:
      summary: Actualiza un registro de progreso
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgresoInput'
      responses:
        '200':
          description: Progreso actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progreso'
    delete:
      summary: Elimina un registro de progreso
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Progreso eliminado
  /progreso/usuario/{usuario_id}:
    get:
      summary: Obtiene todo el progreso de un usuario
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de progresos del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progreso'
  /progreso/usuario/{usuario_id}/rutina/{rutina_id}:
    get:
      summary: Obtiene progreso de un usuario en una rutina específica
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: integer
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Progreso en la rutina
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progreso'
  /progreso/usuario/{usuario_id}/ejercicio/{ejercicio_id}:
    get:
      summary: Obtiene progreso de un usuario en un ejercicio específico
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Progreso en el ejercicio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progreso'
  /progreso/usuario/{usuario_id}/rutina/{rutina_id}/ejercicio/{ejercicio_id}:
    get:
      summary: Obtiene progreso específico en ejercicio dentro de rutina
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: integer
        - in: path
          name: rutina_id
          required: true
          schema:
            type: integer
        - in: path
          name: ejercicio_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Progreso detallado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progreso'
components:
  schemas:
...
    Rutina:
      type: object
      properties:
        id:
          type: integer
          example: 5
        usuario_id:
          type: integer
          example: 1
        entrenador_id:
          type: integer
          example: 2
        descripcion:
          type: string
          example: Rutina de fuerza inicial
      required: [id, usuario_id, entrenador_id, descripcion]
    RutinaInput:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
        entrenador_id:
          type: integer
          example: 2
        descripcion:
          type: string
          example: Rutina de definición
      required: [usuario_id, entrenador_id, descripcion]

    # ======== EJERCICIO ========
    Ejercicio:
      type: object
      properties:
        id:
          type: integer
          example: 8
        nombre:
          type: string
          example: Sentadillas
        tipo:
          type: string
          enum: [pecho, abdominales, piernas, espalda, triceps, biceps, hombro, gluteo]
          example: piernas
        descripcion:
          type: string
          example: Ejercicio para fortalecer piernas
        url_imagen:
          type: string
          example: https://ejemplo.com/sentadilla.jpg
        url_video:
          type: string
          example: https://ejemplo.com/sentadilla.mp4
      required: [id, nombre, tipo]
    EjercicioInput:
      type: object
      properties:
        nombre:
          type: string
          example: Flexiones
        tipo:
          type: string
          enum: [pecho, abdominales, piernas, espalda, triceps, biceps, hombro, gluteo]
          example: pecho
        descripcion:
          type: string
          example: Ejercicio básico para pecho
        url_imagen:
          type: string
          example: https://ejemplo.com/flexiones.jpg
        url_video:
          type: string
          example: https://ejemplo.com/flexiones.mp4
      required: [nombre, tipo]

    # ======== RUTINA_EJERCICIO ========
    RutinaEjercicio:
      type: object
      properties:
        rutina_id:
          type: integer
          example: 5
        ejercicio_id:
          type: integer
          example: 8
        series:
          type: integer
          example: 4
        repeticiones:
          type: integer
          example: 12
        peso_ejercicio:
          type: integer
          example: 20
      required: [rutina_id, ejercicio_id, series, repeticiones]
    RutinaEjercicioInput:
      type: object
      properties:
        ejercicio_id:
          type: integer
          example: 8
        series:
          type: integer
          example: 4
        repeticiones:
          type: integer
          example: 12
        peso_ejercicio:
          type: integer
          example: 20
      required: [ejercicio_id, series, repeticiones]

    # ======== PROGRESO ========
    Progreso:
      type: object
      properties:
        id:
          type: integer
          example: 25
        usuario_id:
          type: integer
          example: 1
        rutina_id:
          type: integer
          example: 5
        ejercicio_id:
          type: integer
          example: 8
        fecha:
          type: string
          format: date
          example: '2023-06-10'
        series_completadas:
          type: integer
          example: 4
        repeticiones_completadas:
          type: integer
          example: 10
        peso_utilizado:
          type: number
          format: float
          example: 22.5
        notas:
          type: string
          example: Última serie con ayuda
      required: [id, usuario_id, rutina_id, ejercicio_id, fecha, series_completadas, repeticiones_completadas]
    ProgresoInput:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
        rutina_id:
          type: integer
          example: 5
        ejercicio_id:
          type: integer
          example: 8
        series_completadas:
          type: integer
          example: 4
        repeticiones_completadas:
          type: integer
          example: 10
        peso_utilizado:
          type: number
          format: float
          example: 22.5
        notas:
          type: string
          example: Última serie con ayuda
      required: [usuario_id, rutina_id, ejercicio_id, series_completadas, repeticiones_completadas]